backend/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
audioop-lts==0.2.2
audioread==3.1.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
fastapi==0.121.2
flatbuffers==25.9.23
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httplib2==0.31.0
idna==3.11
joblib==1.5.2
keras==3.12.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.2
namex==0.1.0
numba==0.62.1
numpy==2.3.4
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.49.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.2.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xgboost==3.1.1
